requirements.txt
aniso8601==8.0.0
attrs==20.2.0
blueprint==3.4.2
click==7.1.2
contextlib2==0.6.0.post1
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
flask-restplus==0.13.0
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.21
protobuf==3.13.0
pyrsistent==0.17.3
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
visitor==0.1.3
validate-docbr==1.8.1
Werkzeug==0.16.1
zipp==3.2.0
gunicorn==20.0.4
